server/tests: require expunge for all tests
